fix version overlay pixels on Forge

+++ forge/src/main/java/net/justmili/leftforgotten/gui/forge/VersionOverlay.java
@@ -1,11 +1,10 @@
 package net.justmili.leftforgotten.gui.forge;
 
+import java.util.Random;
+
 import com.mojang.blaze3d.vertex.PoseStack;
 import net.justmili.leftforgotten.LeftForgotten;
 import net.justmili.leftforgotten.init.LFResources;
-import net.minecraft.client.Minecraft;
-import net.minecraft.client.gui.GuiGraphics;
-import net.minecraft.network.chat.Component;
 import net.minecraftforge.api.distmarker.Dist;
 import net.minecraftforge.client.event.RenderGuiEvent;
 import net.minecraftforge.event.TickEvent;
@@ -13,7 +12,9 @@ import net.minecraftforge.eventbus.api.EventPriority;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.common.Mod;
 
-import java.util.Random;
+import net.minecraft.client.Minecraft;
+import net.minecraft.client.gui.GuiGraphics;
+import net.minecraft.network.chat.Component;
 
 @Mod.EventBusSubscriber(modid = LeftForgotten.MOD_ID, value = Dist.CLIENT)
 public class VersionOverlay {
@@ -69,7 +70,7 @@ public class VersionOverlay {
 
         pose.pushPose();
         pose.scale(1f / guiScaleFactor, 1f / guiScaleFactor, 1f);
-        pose.scale(userScale, userScale, 1f);
+        pose.scale((int) userScale, (int) userScale, 1f);
 
         int x = 6;
         int y = 6;
